[develop] cinema Api requests creation

+++ app/src/main/java/ru/greatdevelopers/android_application/data/dao/CountryDao.kt
@@ -19,5 +19,5 @@ interface CountryDao {
     suspend fun getAllCountry(): List<Country>
 
     @Query("SELECT * FROM Country WHERE country_id == :id")
-    suspend fun getCountryById(id: Int): Country?
+    suspend fun getCountryById(id: Long): Country?
 }
