+++ app/src/main/java/ru/greatdevelopers/android_application/ui/mainscreen/adapters/FilmListItem.kt
@@ -1,4 +1,4 @@
 package ru.greatdevelopers.android_application.ui.mainscreen.adapters
 
-class FilmListItem(var film_id: Int, var film_name: String, var genre_name: String, var poster: String, var rating: Float) {
+class FilmListItem(var film_id: Long, var film_name: String, var genre_name: String, var poster: String, var rating: Float) {
 }
